# swaggerhub/src/IO.Swagger/Controllers/ConfigApi.cs
/*
 * ApiPayment API
 *
 * API for handling payments, subscriptions, invoices, and payment methods for ApiPayment services.
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ConfigApiController : ControllerBase
    { 
        /// <summary>
        /// Get configuration names
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/Config/Names")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigNames")]
        public virtual IActionResult GetConfigNames()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get configuration names by environment
        /// </summary>
        /// <param name="environment"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/Config/Names/{environment}")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigNamesByEnvironment")]
        public virtual IActionResult GetConfigNamesByEnvironment([FromRoute][Required]string environment)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get configuration names in CURL format
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/Config/Names/CURL")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigNamesCurl")]
        public virtual IActionResult GetConfigNamesCurl()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get configuration names in CURL format by environment
        /// </summary>
        /// <param name="environment"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/Config/Names/CURL/{environment}")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigNamesCurlByEnvironment")]
        public virtual IActionResult GetConfigNamesCurlByEnvironment([FromRoute][Required]string environment)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Reload configuration
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/Config/Reload")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigReload")]
        public virtual IActionResult GetConfigReload()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}