# swaggerhub/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ApiPayment API",
    "description" : "API for handling payments, subscriptions, invoices, and payment methods for ApiPayment services.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/Config/Names/{environment}" : {
      "get" : {
        "tags" : [ "Config" ],
        "summary" : "Get configuration names by environment",
        "operationId" : "getConfigNamesByEnvironment",
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
...
    "/Config/Names/CURL/{environment}" : {
      "get" : {
        "tags" : [ "Config" ],
        "summary" : "Get configuration names in CURL format by environment",
        "operationId" : "getConfigNamesCurlByEnvironment",
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
...
  },
...
}